# backend/node/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GymPoint API
  version: 1.0.0
  description: API del sistema GymPoint para gestión de gimnasios, rutinas, entrenamientos y gamificación.
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Datos inválidos o request malformada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Datos inválidos
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso no encontrado
...
  parameters:
...
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador único del recurso
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      additionalProperties: false
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo
          example: Ha ocurrido un error
...
    ExerciseInRoutine:
      type: object
      properties:
        id_exercise:
          type: integer
        exercise_name:
          type: string
        muscular_group:
          type: string
        series:
          type: integer
        reps:
          type: string
        order:
          type: integer
        id_routine_day:
          type: integer
...
    Routine:
      type: object
      properties:
        id_routine:
          type: integer
        routine_name:
          type: string
        description:
          type: string
        created_by:
          type: integer
        is_template:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id_routine
        - routine_name
    RoutineWithDetails:
      allOf:
        - $ref: '#/components/schemas/Routine'
        - type: object
          properties:
            creator:
              type: object
              properties:
                id_user_profile:
                  type: integer
                name:
                  type: string
                lastname:
                  type: string
            days:
              type: array
              items:
                $ref: '#/components/schemas/RoutineDay'
            exercises:
              type: array
              items:
                $ref: '#/components/schemas/ExerciseInRoutine'
    RoutineDay:
      type: object
      properties:
        id_routine_day:
          type: integer
        id_routine:
          type: integer
        day_number:
          type: integer
        title:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id_routine_day
        - id_routine
        - day_number
...
    AssignRoutineRequest:
      type: object
      additionalProperties: false
      properties:
        id_routine:
          type: integer
        start_date:
          type: string
          format: date-time
      required:
        - id_routine
...
    UserRoutine:
      type: object
      properties:
        id_user_routine:
          type: integer
        id_user:
          type: integer
        id_routine:
          type: integer
        start_date:
          type: string
          format: date-time
        finish_date:
          type: string
          format: date-time
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        routine:
          $ref: '#/components/schemas/RoutineWithDetails'
      required:
        - id_user_routine
        - id_user
        - id_routine
        - start_date
        - active
    UserRoutineCounts:
      type: object
      properties:
        total_owned:
          type: integer
        imported_count:
          type: integer
        created_count:
          type: integer
...
paths:
...
  /api/user-routines/assign:
    post:
      operationId: assignRoutineToUser
      tags:
        - User Routines
      summary: Asignar rutina a usuario
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoutineRequest'
      responses:
        '201':
          description: Rutina asignada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserRoutine'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/user-routines/active:
    get:
      operationId: getActiveRoutine
      tags:
        - User Routines
      summary: Obtener rutina activa
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Rutina activa
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserRoutine'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/user-routines/active/details:
    get:
      operationId: getActiveRoutineWithExercises
      tags:
        - User Routines
      summary: Obtener rutina activa con ejercicios
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Rutina activa con ejercicios
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/RoutineWithDetails'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/user-routines/counts:
    get:
      operationId: getUserRoutineCounts
      tags:
        - User Routines
      summary: Obtener contadores de rutinas
      description: Retorna total, importadas y creadas
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Contadores de rutinas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserRoutineCounts'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/user-routines/end:
    post:
      operationId: endUserRoutine
      tags:
        - User Routines
      summary: Finalizar rutina activa
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Rutina finalizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserRoutine'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/user-routines/{id}:
    get:
      operationId: getUserRoutineById
      tags:
        - User Routines
      summary: Obtener rutina de usuario por ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Rutina de usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserRoutine'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/user-routines/{id}/deactivate:
    put:
      operationId: deactivateUserRoutine
      tags:
        - User Routines
      summary: Desactivar rutina de usuario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Rutina desactivada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserRoutine'
        '404':
          $ref: '#/components/responses/NotFound'
...
security:
  - bearerAuth: []